+++ src/simmate/database/core/table.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import inspect
+import logging
 import textwrap
 import urllib
 from functools import cache
@@ -16,9 +17,10 @@ from django.http import HttpRequest, JsonResponse, QueryDict
 from django.shortcuts import get_object_or_404
 from django.urls import resolve, reverse
 from django.utils.module_loading import import_string
+from rich.progress import track
 
 from simmate.config import settings
-from simmate.database.utils import check_db_conn
+from simmate.database.utils import batch_bulk_create, check_db_conn
 from simmate.utils import get_attributes_doc
 
 from .archive import ArchiveMixin
@@ -1011,3 +1013,20 @@ class DatabaseTable(models.Model, ArchiveMixin):
         if queryset is None:
             queryset = cls.objects
         return queryset.to_dataframe()  # default to just raw data
+
+    # -------------------------------------------------------------------------
+    # Methods for bulk-loading data from external sources
+    # -------------------------------------------------------------------------
+
+    @classmethod
+    def load_source_data(cls, **kwargs):
+        """
+        Loads all data from the source into the local Simmate database.
+
+        This method should be implemented as a generator that yields initialized
+        database objects. You should also use the `@batch_bulk_create` decorator
+        to handle saving the objects in batches.
+        """
+        raise NotImplementedError(
+            "The `load_source_data` method has not been implemented for this table."
+        )