+++ docs/change_log.md
@@ -77,6 +77,8 @@ There is one significant exception to the above rules -- `MAJOR=0` releases. Any
 - added a `@schedule` decorator for easier periodic task registration
 - added web components and templates for the pdb and chembl apps
 - added crystalline dataset views to the Data Explorer, including 3D structure views, exports, and search forms
+- added `@batch_bulk_create` decorator to help simplify `load_source_data` methods
+- updated `simmate database download` to be dynamic across all apps, support a `--source` option, and respect custom load orders
 
 **Refactors**